requirements.txt
youtube-transcript-api
langchain-community
langchain-openai
langchain-core
langchain-text-splitters
faiss-cpu
tiktoken
python-dotenv
cmake
pyarrow
streamlit
